Don't use auto-generated classes for selectors
The classes used in Meet are auto-generated and are thus brittle and not ideal to use for selectors. This update to reverseNames instead relies on the structure of the page to target the participant names. Because the classes that were previously in use as selectors are no longer present,

I'm unsure if this targeted the names in the videos themselves or just simply the list of participants. If the former, this update won't fix anything; if the latter, this update should be a decent fix for now (until Google changes the document structure).